// src/features/home/calendar/interfaces.ts
import type { TCalendarType } from '@/features/home/calendar/types';

export interface IUser {
  id: string;
  name: string;
  picturePath: string | null;
}

export interface IEvent {
  id: number;
  startDate: string;
  endDate: string;
  title: string;
  description: string;
  vacationType?: string;
  user: IUser;
  type: TCalendarType;
}

export interface ICalendarCell {
  day: number;
  currentMonth: boolean;
  date: Date;
}